+++ src/components/InstanceFilters/InstanceFilters.js
@@ -2,18 +2,16 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage, useIntl } from 'react-intl';
 import _ from 'lodash';
-import moment from 'moment';
 
 import {
   Accordion,
   AccordionSet,
   FilterAccordionHeader,
 } from '@folio/stripes/components';
-import { DateRangeFilter } from '@folio/stripes/smart-components';
-import { useStripes } from '@folio/stripes/core';
 
 import TagsFilter from '../TagsFilter';
 import CheckboxFacet from '../CheckboxFacet';
+import DateRangeFilter from '../DateRangeFilter';
 import {
   getSourceOptions,
   getSuppressedOptions,
@@ -34,6 +32,7 @@ import {
   retrieveDatesFromDateRangeFilterString,
 } from '../../utils';
 
+
 const InstanceFilters = props => {
   const {
     activeFilters,
@@ -51,7 +50,6 @@ const InstanceFilters = props => {
   } = props;
 
   const intl = useIntl();
-  const { timezone } = useStripes();
 
   const segmentAccordions = {
     [FACETS.EFFECTIVE_LOCATION]: false,
@@ -166,12 +164,6 @@ const InstanceFilters = props => {
     getNewRecords,
     props.data
   );
-
-  const onDateRangeChange = (filterData) => {
-    moment.tz.setDefault(timezone);
-    onChange(filterData);
-    moment.tz.setDefault();
-  };
 
   return (
     <AccordionSet accordionStatus={accordions} onToggle={onToggleSection}>
@@ -363,7 +355,7 @@ const InstanceFilters = props => {
           name={FACETS.CREATED_DATE}
           dateFormat={DATE_FORMAT}
           selectedValues={retrieveDatesFromDateRangeFilterString(activeFilters[FACETS.CREATED_DATE]?.[0])}
-          onChange={onDateRangeChange}
+          onChange={onChange}
           makeFilterString={makeDateRangeFilterString}
         />
       </Accordion>
@@ -380,7 +372,7 @@ const InstanceFilters = props => {
           name={FACETS.UPDATED_DATE}
           dateFormat={DATE_FORMAT}
           selectedValues={retrieveDatesFromDateRangeFilterString(activeFilters[FACETS.UPDATED_DATE]?.[0])}
-          onChange={onDateRangeChange}
+          onChange={onChange}
           makeFilterString={makeDateRangeFilterString}
         />
       </Accordion>
